WIP: Refactor adapter to act via readers
